split sources to own file
